backend/requirements.txt
asgiref==3.8.1
Django==4.2.20
django-cors-headers==4.7.0
django-environ==0.12.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
phonenumbers==9.0.1
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.12.2
